+++ src/main/java/com/todorian/_core/jwt/JWTTokenProvider.java
@@ -1,6 +1,6 @@
 package com.todorian._core.jwt;
 
-import com.todorian.member.dto.MemberResponseDTO;
+import com.todorian.member.command.application.dto.MemberResponseDTO;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
