+++ src/main/java/com/todorian/member/command/domain/model/Authority.java
@@ -1,4 +1,4 @@
-package com.todorian.member.domain;
+package com.todorian.member.command.domain.model;
 
 public enum Authority {
     ADMIN, USER, GUEST, NONE
