+++ src/main/java/com/todorian/member/command/domain/model/Status.java
@@ -1,4 +1,4 @@
-package com.todorian.member.domain;
+package com.todorian.member.command.domain.model;
 
 public enum Status {
     ACTIVE, DORMANT, DEACTIVATED, SUSPENDED
