+++ src/test/java/com/todorian/member/service/MemberAuthServiceTest.java
@@ -1,5 +1,6 @@
 package com.todorian.member.service;
 
+import com.todorian.member.command.application.service.MemberAuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
